feat: Create the Logo, Header, Footer and Button components.
Create the Logo component to be used in both Header and Footer components, but haven't created the logic for the display size change, which will affect the rendered size of the logo. Installed the Phosphor Icons package. I decided to use the Tailwind Variants to create the Button component, so it would be easier and way more clear to change the style of the button in different places. In the Header component, given the fact that it's a small component, I decided to implement the composition pattern inside the own component instead of creating two new components. Seems to be everything ok.