feat: Create the initial structure for the Section component.
Create the initial structure of the Section component and create the BenefitsCard component as well. Create the data folder which will hold the TypeScript files with the data that will be provided for the Sections and pages of the project. The Section component will be refactored to implement the composition pattern.